requirements.txt
# Core LangChain packages
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Model connectors
langchain-openai>=0.1.0   # optional OpenAI fallback
langchain-groq>=0.1.0     # Groq models like llama3

# Environment variable loader
python-dotenv>=1.0.0

# Streamlit UI
streamlit>=1.40.0

# Vector store for Q&A
faiss-cpu>=1.7.4

# Document parsing
unstructured>=0.12.0
pdfminer.six>=20221105
python-docx>=1.1.0
xlrd>=2.0.1
openpyxl>=3.1.0   # modern Excel support
pandas>=2.0.0

# Tokenization
tiktoken>=0.7.0   # required for LangChain text splitting

# Numerical computing
numpy>=1.26.0     # used by FAISS and loaders

# Optional helpers
requests>=2.31.0
beautifulsoup4>=4.12.2
pillow>=10.0.0    # image handling support
python-magic>=0.4.27  # file type detection